Move Fragments to ui package
