+++ app/src/main/java/com/shapeapp/shape/ui/ReceivedImageFragmentViewModel.kt
@@ -5,7 +5,6 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.shapeapp.shape.data.Card
-import com.shapeapp.shape.fragments.ReceivedImageFragment
 
 /**
  * [ViewModel] for [ReceivedImageFragment]
