+++ app/src/main/java/com/shapeapp/shape/ui/TakePhotoFragmentViewModel.kt
@@ -4,7 +4,6 @@ import android.net.Uri
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import com.shapeapp.shape.fragments.TakePhotoFragment
 
 /**
  * [ViewModel] for [TakePhotoFragment]
