+++ app/src/main/res/layout/fragment_camera.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.CameraFragment">
+    tools:context=".ui.CameraFragment">
 
     <io.fotoapparat.view.CameraView
         android:id="@+id/preview_cameraview"
