+++ app/src/main/res/layout/fragment_initial_backend_connection.xml
@@ -5,7 +5,7 @@
     android:id="@+id/root_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.InitialBackendConnectionFragment">
+    tools:context=".ui.InitialBackendConnectionFragment">
 
     <TextView
         android:id="@+id/title_textview"
