+++ app/src/main/res/layout/fragment_profile.xml
@@ -5,7 +5,7 @@
     android:id="@+id/root_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.ProfileFragment">
+    tools:context=".ui.ProfileFragment">
 
 
     <androidx.constraintlayout.widget.ConstraintLayout
