+++ app/src/main/res/layout/fragment_public.xml
@@ -5,7 +5,7 @@
     android:id="@+id/root_coordinatorlayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.PublicFragment">
+    tools:context=".ui.PublicFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/root_layout"
