+++ app/src/main/res/layout/fragment_received_image.xml
@@ -5,7 +5,7 @@
     android:id="@+id/root_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.ReceivedImageFragment">
+    tools:context=".ui.ReceivedImageFragment">
 
     <ImageView
         android:id="@+id/background_imageview"
